ready to publish version 0.7.0
